computer-vision/human-segmentation-coco/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.7.0
pycocotools>=2.0.0
matplotlib>=3.5.0
numpy>=1.21.0
tqdm>=4.64.0
Pillow>=9.0.0
scipy>=1.9.0
PyYAML>=6.0
scikit-learn>=1.3.0